requirements.txt
tensorflow
matplotlib
huggingface-hub
flask
pyngrok
pillow
opencv-python
python-dotenv
